Create test cases
This commit automatically downloads GTest during the build phase to
compile it as a target, so we can use CTest.

Implemented a simple unit test for the polya contagion.

Added some methods to EigenTypes so we can create matrices more easily.
This will be improved in future commits, so we can use more natural
commands than what Eigen allows for.